review + refactoring
- separated dynamoplus from aws
